Workaround fletchc issue which makes it unable to compile code importing "dart:isolate"
Since the Dart sdk sources are used for:
 * analyzing code for warnings/errors
 * running fletchc itself
 * compiling a fletchc-deltablue compiler (i.e. fletchc compiles itself)

In particular the last point involves being able to compile all of fletchc code to bytecodes. The fletchc compiler cannot do that if it sees imports of "dart:isolate" and we workaround it for that reason.

R=ahe@google.com

Review URL: https://codereview.chromium.org/1600893003 .